Undo renaming since the module is no longer user facing
