+++ strawberryfields/apps/train/__init__.py
@@ -80,7 +80,7 @@ function, gradients can be computed efficiently, leading to fast training.
 """
 from strawberryfields.apps.train.cost import KL, Stochastic
 from strawberryfields.apps.train.embed import Exp, ExpFeatures
-from strawberryfields.apps.train.vgbs import (
+from strawberryfields.apps.train.param import (
     VGBS,
     A_to_cov,
     prob_click,
